High-Level Interface

Hash Functions

CryptographicHashFunctions.blake2b_160 — Method

blake2b_160(data[; provider, kwargs...])

Return the digest of data computed with the BLAKE2B_160 algorithm.

CryptographicHashFunctions.blake2b_256 — Method

blake2b_256(data[; provider, kwargs...])

Return the digest of data computed with the BLAKE2B_256 algorithm.

CryptographicHashFunctions.blake2b_384 — Method

blake2b_384(data[; provider, kwargs...])

Return the digest of data computed with the BLAKE2B_384 algorithm.

CryptographicHashFunctions.blake2b_512 — Method

blake2b_512(data[; provider, kwargs...])

Return the digest of data computed with the BLAKE2B_512 algorithm.

CryptographicHashFunctions.blake2s_128 — Method

blake2s_128(data[; provider, kwargs...])

Return the digest of data computed with the BLAKE2S_128 algorithm.

CryptographicHashFunctions.blake2s_160 — Method

blake2s_160(data[; provider, kwargs...])

Return the digest of data computed with the BLAKE2S_160 algorithm.

CryptographicHashFunctions.blake2s_224 — Method

blake2s_224(data[; provider, kwargs...])

Return the digest of data computed with the BLAKE2S_224 algorithm.

CryptographicHashFunctions.blake2s_256 — Method

blake2s_256(data[; provider, kwargs...])

Return the digest of data computed with the BLAKE2S_256 algorithm.

CryptographicHashFunctions.gost94 — Method

gost94(data[; provider, kwargs...])

Return the digest of data computed with the GOST94 algorithm.

CryptographicHashFunctions.gost94cp — Method

gost94cp(data[; provider, kwargs...])

Return the digest of data computed with the GOST94CP algorithm.

CryptographicHashFunctions.md2 — Method

md2(data[; provider, kwargs...])

Return the digest of data computed with the MD2 algorithm.

CryptographicHashFunctions.md4 — Method

md4(data[; provider, kwargs...])

Return the digest of data computed with the MD4 algorithm.

CryptographicHashFunctions.md5 — Method

md5(data[; provider, kwargs...])

Return the digest of data computed with the MD5 algorithm.

CryptographicHashFunctions.ripemd160 — Method

ripemd160(data[; provider, kwargs...])

Return the digest of data computed with the RIPEMD160 algorithm.

CryptographicHashFunctions.sha1 — Method

sha1(data[; provider, kwargs...])

Return the digest of data computed with the SHA1 algorithm.

CryptographicHashFunctions.sha224 — Method

sha224(data[; provider, kwargs...])

Return the digest of data computed with the SHA224 algorithm.

CryptographicHashFunctions.sha256 — Method

sha256(data[; provider, kwargs...])

Return the digest of data computed with the SHA256 algorithm.

CryptographicHashFunctions.sha384 — Method

sha384(data[; provider, kwargs...])

Return the digest of data computed with the SHA384 algorithm.

CryptographicHashFunctions.sha3_224 — Method

sha3_224(data[; provider, kwargs...])

Return the digest of data computed with the SHA3_224 algorithm.

CryptographicHashFunctions.sha3_256 — Method

sha3_256(data[; provider, kwargs...])

Return the digest of data computed with the SHA3_256 algorithm.

CryptographicHashFunctions.sha3_384 — Method

sha3_384(data[; provider, kwargs...])

Return the digest of data computed with the SHA3_384 algorithm.

CryptographicHashFunctions.sha3_512 — Method

sha3_512(data[; provider, kwargs...])

Return the digest of data computed with the SHA3_512 algorithm.

CryptographicHashFunctions.sha512 — Method

sha512(data[; provider, kwargs...])

Return the digest of data computed with the SHA512 algorithm.

CryptographicHashFunctions.sha512_224 — Method

sha512_224(data[; provider, kwargs...])

Return the digest of data computed with the SHA512_224 algorithm.

CryptographicHashFunctions.sha512_256 — Method

sha512_256(data[; provider, kwargs...])

Return the digest of data computed with the SHA512_256 algorithm.

CryptographicHashFunctions.sm3 — Method

sm3(data[; provider, kwargs...])

Return the digest of data computed with the SM3 algorithm.

CryptographicHashFunctions.stribog256 — Method

stribog256(data[; provider, kwargs...])

Return the digest of data computed with the STRIBOG256 algorithm.

CryptographicHashFunctions.stribog512 — Method

stribog512(data[; provider, kwargs...])

Return the digest of data computed with the STRIBOG512 algorithm.

CryptographicHashFunctions.tiger — Method

tiger(data[; provider, kwargs...])

Return the digest of data computed with the TIGER algorithm.

CryptographicHashFunctions.tiger1 — Method

tiger1(data[; provider, kwargs...])

Return the digest of data computed with the TIGER1 algorithm.

CryptographicHashFunctions.tiger2 — Method

tiger2(data[; provider, kwargs...])

Return the digest of data computed with the TIGER2 algorithm.

CryptographicHashFunctions.whirlpool — Method

whirlpool(data[; provider, kwargs...])

Return the digest of data computed with the WHIRLPOOL algorithm.

HMAC Functions

CryptographicHashFunctions.hmac_blake2b_160 — Method

hmac_blake2b_160(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the BLAKE2B_160 algorithm.

CryptographicHashFunctions.hmac_blake2b_256 — Method

hmac_blake2b_256(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the BLAKE2B_256 algorithm.

CryptographicHashFunctions.hmac_blake2b_384 — Method

hmac_blake2b_384(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the BLAKE2B_384 algorithm.

CryptographicHashFunctions.hmac_blake2b_512 — Method

hmac_blake2b_512(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the BLAKE2B_512 algorithm.

CryptographicHashFunctions.hmac_blake2s_128 — Method

hmac_blake2s_128(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the BLAKE2S_128 algorithm.

CryptographicHashFunctions.hmac_blake2s_160 — Method

hmac_blake2s_160(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the BLAKE2S_160 algorithm.

CryptographicHashFunctions.hmac_blake2s_224 — Method

hmac_blake2s_224(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the BLAKE2S_224 algorithm.

CryptographicHashFunctions.hmac_blake2s_256 — Method

hmac_blake2s_256(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the BLAKE2S_256 algorithm.

CryptographicHashFunctions.hmac_gost94 — Method

hmac_gost94(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the GOST94 algorithm.

CryptographicHashFunctions.hmac_gost94cp — Method

hmac_gost94cp(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the GOST94CP algorithm.

CryptographicHashFunctions.hmac_md2 — Method

hmac_md2(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the MD2 algorithm.

CryptographicHashFunctions.hmac_md4 — Method

hmac_md4(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the MD4 algorithm.

CryptographicHashFunctions.hmac_md5 — Method

hmac_md5(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the MD5 algorithm.

CryptographicHashFunctions.hmac_ripemd160 — Method

hmac_ripemd160(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the RIPEMD160 algorithm.

CryptographicHashFunctions.hmac_sha1 — Method

hmac_sha1(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA1 algorithm.

CryptographicHashFunctions.hmac_sha224 — Method

hmac_sha224(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA224 algorithm.

CryptographicHashFunctions.hmac_sha256 — Method

hmac_sha256(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA256 algorithm.

CryptographicHashFunctions.hmac_sha384 — Method

hmac_sha384(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA384 algorithm.

CryptographicHashFunctions.hmac_sha3_224 — Method

hmac_sha3_224(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA3_224 algorithm.

CryptographicHashFunctions.hmac_sha3_256 — Method

hmac_sha3_256(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA3_256 algorithm.

CryptographicHashFunctions.hmac_sha3_384 — Method

hmac_sha3_384(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA3_384 algorithm.

CryptographicHashFunctions.hmac_sha3_512 — Method

hmac_sha3_512(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA3_512 algorithm.

CryptographicHashFunctions.hmac_sha512 — Method

hmac_sha512(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA512 algorithm.

CryptographicHashFunctions.hmac_sha512_224 — Method

hmac_sha512_224(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA512_224 algorithm.

CryptographicHashFunctions.hmac_sha512_256 — Method

hmac_sha512_256(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SHA512_256 algorithm.

CryptographicHashFunctions.hmac_sm3 — Method

hmac_sm3(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the SM3 algorithm.

CryptographicHashFunctions.hmac_stribog256 — Method

hmac_stribog256(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the STRIBOG256 algorithm.

CryptographicHashFunctions.hmac_stribog512 — Method

hmac_stribog512(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the STRIBOG512 algorithm.

CryptographicHashFunctions.hmac_tiger — Method

hmac_tiger(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the TIGER algorithm.

CryptographicHashFunctions.hmac_tiger1 — Method

hmac_tiger1(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the TIGER1 algorithm.

CryptographicHashFunctions.hmac_tiger2 — Method

hmac_tiger2(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the TIGER2 algorithm.

CryptographicHashFunctions.hmac_whirlpool — Method

hmac_whirlpool(key, data[; provider, kwargs...])

Return the HMAC of data keyed with key computed with the WHIRLPOOL algorithm.

XOF Functions

CryptographicHashFunctions.shake128 — Method

shake128(data, len[; provider, kwargs...])

Return the first len bytes of the digest of data computed with the SHAKE128 algorithm.

CryptographicHashFunctions.shake128 — Method

shake128(data[; provider, kwargs...])

Return a hash context initialized with data for the SHAKE128 algorithm.

CryptographicHashFunctions.shake256 — Method

shake256(data, len[; provider, kwargs...])

Return the first len bytes of the digest of data computed with the SHAKE256 algorithm.

CryptographicHashFunctions.shake256 — Method

shake256(data[; provider, kwargs...])

Return a hash context initialized with data for the SHAKE256 algorithm.